+++ pkg/reviewapps/reviewapps.go
@@ -1,14 +1,10 @@
 package reviewapps
 
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"strings"
 	"tuber/pkg/k8s"
-
-	"google.golang.org/api/cloudbuild/v1"
-	"google.golang.org/api/option"
 )
 
 // NewReviewAppSetup replicates a namespace and its roles, rolebindings, and opaque secrets after removing their non-generic metadata.
@@ -110,71 +106,4 @@ func duplicateResource(resource []byte, sourceApp string, reviewApp string) ([]b
 		return nil, err
 	}
 	return genericized, nil
-}
-
-const tuberReposConfig = "tuber-repos"
-const tuberReviewTriggersConfig = "tuber-review-triggers"
-
-func CreateAndRunTrigger(ctx context.Context, creds []byte, sourceRepo string, project string, targetAppName string, branch string) error {
-	config, err := k8s.GetConfig(tuberReposConfig, "tuber", "configmap")
-	if err != nil {
-		return err
-	}
-	var cloudSourceRepo string
-	for k, v := range config.Data {
-		if v == sourceRepo {
-			cloudSourceRepo = k
-		}
-	}
-	if cloudSourceRepo == "" {
-		return fmt.Errorf("source repo not present in tuber-repos")
-	}
-	cloudbuildService, err := cloudbuild.NewService(ctx, option.WithCredentialsJSON(creds))
-	if err != nil {
-		return err
-	}
-	service := cloudbuild.NewProjectsTriggersService(cloudbuildService)
-	triggerTemplate := cloudbuild.RepoSource{
-		BranchName: branch,
-		ProjectId:  project,
-		RepoName:   cloudSourceRepo,
-	}
-	buildTrigger := cloudbuild.BuildTrigger{
-		Description:     "created by tuber",
-		Filename:        "cloudbuild.yaml",
-		Name:            "review-app-for-" + targetAppName,
-		TriggerTemplate: &triggerTemplate,
-	}
-	triggerCreateCall := service.Create(project, &buildTrigger)
-	triggerCreateResult, err := triggerCreateCall.Do()
-	if err != nil {
-		return err
-	}
-	err = k8s.PatchConfigMap(tuberReviewTriggersConfig, "tuber", targetAppName, triggerCreateResult.Id)
-	if err != nil {
-		deleteErr := deleteReviewAppTrigger(service, project, triggerCreateResult.Id, targetAppName)
-		if deleteErr != nil {
-			return fmt.Errorf(err.Error() + deleteErr.Error())
-		}
-		return err
-	}
-	triggerRunCall := service.Run(project, triggerCreateResult.Id, &triggerTemplate)
-	_, err = triggerRunCall.Do()
-	if err != nil {
-		deleteErr := deleteReviewAppTrigger(service, project, triggerCreateResult.Id, targetAppName)
-		if deleteErr != nil {
-			return fmt.Errorf(err.Error() + deleteErr.Error())
-		}
-		return err
-	}
-	return nil
-}
-
-func deleteReviewAppTrigger(service *cloudbuild.ProjectsTriggersService, project string, triggerId string, appName string) error {
-	deleteCall := service.Delete(project, triggerId)
-	_, err := deleteCall.Do()
-	if err != nil {
-		return err
-	}
-	return k8s.RemoveConfigMapEntry(tuberReviewTriggersConfig, "tuber", appName)
 }
